+++ dataprovider/basketballreference/nba/basic_box_score_stats.go
@@ -1,10 +1,8 @@
 package nba
 
 import (
-	"context"
 	"fmt"
 	"log"
-	"strings"
 	"sync"
 	"time"
 	"unicode"
@@ -12,24 +10,15 @@ import (
 	"github.com/lightning-dabbler/sportscrape/dataprovider/basketballreference"
 	"github.com/lightning-dabbler/sportscrape/dataprovider/basketballreference/nba/model"
 	"github.com/lightning-dabbler/sportscrape/util"
+	"github.com/lightning-dabbler/sportscrape/util/request"
 
 	"github.com/PuerkitoBio/goquery"
-	"github.com/chromedp/cdproto/network"
-	"github.com/chromedp/chromedp"
 )
 
 const (
-	// https://www.basketball-reference.com/boxscores/{event_id}.html Selectors for Box Scores
-	basicBoxScoreSelector             = `table[id$='game-basic']`
-	basicBoxScoreContentSelector      = `#content`
-	basicBoxScoreStatsRecordsSelector = `tbody > tr`
-	basicBoxScoreStarterHeaders       = `thead > tr:nth-child(2) th`
-	basicBoxScoreReserveHeaders       = `th`
-	basicBoxScorePlayerSelector       = "th"
-	basicBoxScorePlayerLinkSelector   = basicBoxScorePlayerSelector + " > a"
+	// Selector for Basic Box Score stats tables
+	basicBoxScoreSelector = `table[id$='game-basic']`
 )
-
-type headerValues map[string]struct{}
 
 // basicBoxScoreStarterHeaderValues represents the headers in sequential order for the starter team members
 var basicBoxScoreStarterHeaderValues headerValues = headerValues{
@@ -89,58 +78,27 @@ var basicBoxScoreReservesHeaderValues headerValues = headerValues{
 func getBasicBoxScoreStats(nbaMatchup interface{}) []interface{} {
 	matchup := nbaMatchup.(model.NBAMatchup)
 	url := matchup.BoxScoreLink
-	var basicNBABoxScoreStats []interface{}
-	ctx, cancel := chromedp.NewContext(context.Background())
-	defer cancel()
-	ctx, cancel = context.WithTimeout(ctx, 2*time.Minute)
-	defer cancel()
-	fmt.Println("Scraping Basic Box Score: " + url)
-	var outer string
 	PullTimestamp := time.Now().UTC()
 	start := time.Now().UTC()
-	if err := chromedp.Run(ctx,
-		network.Enable(),
-		network.SetExtraHTTPHeaders(network.Headers(map[string]interface{}{
-			"authority":                 "www.basketball-reference.com",
-			"accept":                    "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9",
-			"accept-language":           "en-US;q=0.8",
-			"cookie":                    "is_live=true; sr_note_box_countdown=57",
-			"if-modified-since":         "Tue, 08 Nov 2022 01:08:31 GMT",
-			"sec-fetch-dest":            "document",
-			"sec-fetch-mode":            "navigate",
-			"sec-fetch-site":            "none",
-			"sec-fetch-user":            "?1",
-			"sec-gpc":                   "1",
-			"upgrade-insecure-requests": "1",
-			"user-agent":                "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/103.0.5060.53 Safari/537.36",
-		})),
-		chromedp.Navigate(url),
-		chromedp.WaitReady(basicBoxScoreContentSelector),
-		chromedp.OuterHTML(basicBoxScoreContentSelector, &outer, chromedp.ByQuery),
-	); err != nil {
-		fmt.Println("ERROR:")
-		fmt.Printf("getBasicBoxScoreStats %s\n", url)
-		log.Fatalln(err)
-	}
-	myReader := strings.NewReader(outer)
-	doc, err := goquery.NewDocumentFromReader(myReader)
+	var basicNBABoxScoreStats []interface{}
+	fmt.Println("Scraping Basic Box Score: " + url)
+	dr := request.NewDocumentRetriever(2 * time.Minute)
+	doc, err := dr.RetrieveDocument(url, networkHeaders, waitReadyBoxScoreContentSelector)
 	if err != nil {
 		log.Fatalln(err)
 	}
 	doc.Find(basicBoxScoreSelector).Each(func(i int, s *goquery.Selection) {
 		var starterHeader string
 		var reserveHeader string
-		s.Find(basicBoxScoreStarterHeaders).Each(func(_ int, s *goquery.Selection) {
-
+		s.Find(boxScoreStarterHeaders).Each(func(_ int, s *goquery.Selection) {
 			starterHeader = util.CleanTextDatum(s.Text())
 			_, ok := basicBoxScoreStarterHeaderValues[starterHeader]
 			if !ok {
 				log.Fatalf("%s is not a valid Starters Header @ %s\n", starterHeader, url)
 			}
-
 		})
 
-		s.Find(basicBoxScoreStatsRecordsSelector).Each(func(j int, s *goquery.Selection) {
+		s.Find(boxScoreStatsRecordsSelector).Each(func(j int, s *goquery.Selection) {
 			var boxScoreStats model.NBABasicBoxScoreStats
 			if j < 5 || j > 5 {
 				boxScoreStats.PullTimestamp = PullTimestamp
@@ -158,106 +116,90 @@ func getBasicBoxScoreStats(nbaMatchup interface{}) []interface{} {
 				} else {
 					boxScoreStats.Starter = false
 				}
-				boxScoreStats.PlayerLink = basketballreference.URL + util.CleanTextDatum(s.Find(basicBoxScorePlayerLinkSelector).AttrOr("href", ""))
-				boxScoreStats.Player = util.CleanTextDatum(s.Find(basicBoxScorePlayerSelector).Text())
-				playerLinkSplit := strings.Split(boxScoreStats.PlayerLink, "/")
-				boxScoreStats.PlayerID = strings.Split(playerLinkSplit[len(playerLinkSplit)-1], ".")[0]
+				boxScoreStats.PlayerLink = basketballreference.URL + util.CleanTextDatum(s.Find(boxScorePlayerLinkSelector).AttrOr("href", ""))
+				boxScoreStats.Player = util.CleanTextDatum(s.Find(boxScorePlayerSelector).Text())
+				playerID, err := extractPlayerID(boxScoreStats.PlayerLink)
+				if err != nil {
+					log.Fatalln(err)
+				}
+				boxScoreStats.PlayerID = playerID
 				minutesPlayed := util.CleanTextDatum(s.Find("td:nth-child(2)").Text())
 				if len(minutesPlayed) > 0 && unicode.IsDigit(rune(minutesPlayed[0])) {
-					minutesPlayedSplit := strings.Split(minutesPlayed, ":")
-					minutes, err := util.TextToInt(minutesPlayedSplit[0])
+					totalMinutes, err := transformMinutesPlayed(minutesPlayed)
 					if err != nil {
-						log.Printf("Could not convert minutes %s to integer\n", minutesPlayedSplit[0])
 						log.Fatalln(err)
 					}
-
-					seconds, err := util.TextToInt(minutesPlayedSplit[1])
-					if err != nil {
-						log.Printf("Could not convert seconds %s to integer\n", minutesPlayedSplit[0])
-						log.Fatalln(err)
-					}
-
-					totalMinutes := float32(minutes) + (float32(seconds) / float32(60))
 					boxScoreStats.MinutesPlayed = totalMinutes
 
 					fieldGoalsMadeText := util.CleanTextDatum(s.Find("td:nth-child(3)").Text())
 					fieldGoalsMade, err := util.TextToInt(fieldGoalsMadeText)
 					if err != nil {
-						log.Printf("Can't convert '%s' for fieldGoalsMadeText to Int\n", fieldGoalsMadeText)
+						log.Println(fmt.Errorf("WARNING: Can't convert '%s' for fieldGoalsMadeText to Int - %w; defaulting to 0.", fieldGoalsMadeText, err))
 						fieldGoalsMade = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 					fieldGoalAttemptsText := util.CleanTextDatum(s.Find("td:nth-child(4)").Text())
 					fieldGoalAttempts, err := util.TextToInt(fieldGoalAttemptsText)
 					if err != nil {
-						log.Printf("Can't convert '%s' for fieldGoalAttemptsText to Int\n", fieldGoalAttemptsText)
+						log.Println(fmt.Errorf("WARNING: Can't convert '%s' for fieldGoalAttemptsText to Int - %w; defaulting to 0.", fieldGoalAttemptsText, err))
 						fieldGoalAttempts = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 					rawFieldGoalPercentage := util.CleanTextDatum(s.Find("td:nth-child(5)").Text())
 					var fieldGoalPercentage float32
 
 					if rawFieldGoalPercentage == "" {
-						fieldGoalPercentage = 0.00
+						fieldGoalPercentage = 0
 					} else {
 						fieldGoalPercentage, err = util.TextToFloat32(rawFieldGoalPercentage)
 						if err != nil {
-							log.Printf("Can't convert '%s' for rawFieldGoalPercentage to Float64\n", rawFieldGoalPercentage)
-							log.Fatalln(err)
+							log.Fatalln(fmt.Errorf("Can't convert '%s' for rawFieldGoalPercentage to Float64 - %w", rawFieldGoalPercentage, err))
 						}
 					}
 
 					threePointsMadeText := util.CleanTextDatum(s.Find("td:nth-child(6)").Text())
 					threePointsMade, err := util.TextToInt(threePointsMadeText)
 					if err != nil {
-						log.Printf("Can't convert '%s' for threePointsMadeText to Int\n", threePointsMadeText)
+						log.Println(fmt.Errorf("WARNING: Can't convert '%s' for threePointsMadeText to Int - %w; defaulting to 0.", threePointsMadeText, err))
 						threePointsMade = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 					threePointAttemptsText := util.CleanTextDatum(s.Find("td:nth-child(7)").Text())
 					threePointAttempts, err := util.TextToInt(threePointAttemptsText)
 					if err != nil {
-						log.Printf("Can't convert '%s' for threePointAttemptsText to Int\n", threePointAttemptsText)
+						log.Println(fmt.Errorf("WARNING: Can't convert '%s' for threePointAttemptsText to Int - %w; defaulting to 0.", threePointAttemptsText, err))
 						threePointsMade = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 					rawthreePointPercentage := util.CleanTextDatum(s.Find("td:nth-child(8)").Text())
 					var threePointPercentage float32
 
 					if rawthreePointPercentage == "" {
-						threePointPercentage = 0.00
+						threePointPercentage = 0
 					} else {
 						threePointPercentage, err = util.TextToFloat32(rawthreePointPercentage)
 						if err != nil {
-							log.Printf("Can't convert '%s' for rawthreePointPercentage to Float64\n", rawthreePointPercentage)
-							log.Fatalln(err)
+							log.Fatalln(fmt.Errorf("Can't convert '%s' for rawthreePointPercentage to Float64 - %w", rawthreePointPercentage, err))
 						}
 					}
 
 					freeThrowsMadeText := util.CleanTextDatum(s.Find("td:nth-child(9)").Text())
 					freeThrowsMade, err := util.TextToInt(freeThrowsMadeText)
 					if err != nil {
-						log.Printf("Can't convert '%s' for freeThrowsMadeText to Int\n", freeThrowsMadeText)
+						log.Println(fmt.Errorf("WARNING: Can't convert '%s' for freeThrowsMadeText to Int - %w; defaulting to 0.", freeThrowsMadeText, err))
 						freeThrowsMade = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 					freeThrowAttemptsText := util.CleanTextDatum(s.Find("td:nth-child(10)").Text())
 					freeThrowAttempts, err := util.TextToInt(freeThrowAttemptsText)
 					if err != nil {
-						log.Printf("Can't convert '%s' for freeThrowAttemptsText to Int\n", freeThrowAttemptsText)
+						log.Println(fmt.Errorf("WARNING: Can't convert '%s' for freeThrowAttemptsText to Int - %w; defaulting to 0.", freeThrowAttemptsText, err))
 						freeThrowAttempts = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 					rawfreeThrowPercentage := util.CleanTextDatum(s.Find("td:nth-child(11)").Text())
 					var freeThrowPercentage float32
 
 					if rawfreeThrowPercentage == "" {
-						freeThrowPercentage = 0.00
+						freeThrowPercentage = 0
 					} else {
 						freeThrowPercentage, err = util.TextToFloat32(rawfreeThrowPercentage)
 						if err != nil {
-							log.Printf("Can't convert '%s' for rawfreeThrowPercentage to Float64\n", rawfreeThrowPercentage)
-							log.Fatalln(err)
+							log.Fatalln(fmt.Errorf("Error: Can't convert '%s' for rawfreeThrowPercentage to Float64 - %w", rawfreeThrowPercentage, err))
 						}
 					}
 					boxScoreStats.FieldGoalsMade = fieldGoalsMade
@@ -273,96 +215,85 @@ func getBasicBoxScoreStats(nbaMatchup interface{}) []interface{} {
 					OffensiveReboundsText := util.CleanTextDatum(s.Find("td:nth-child(12)").Text())
 					boxScoreStats.OffensiveRebounds, err = util.TextToInt(OffensiveReboundsText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for OffensiveReboundsText to Int\n", OffensiveReboundsText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for OffensiveReboundsText to Int - %w; defaulting to 0.", OffensiveReboundsText, err))
 						boxScoreStats.OffensiveRebounds = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					DefensiveReboundsText := util.CleanTextDatum(s.Find("td:nth-child(13)").Text())
 					boxScoreStats.DefensiveRebounds, err = util.TextToInt(DefensiveReboundsText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for DefensiveReboundsText to Int\n", DefensiveReboundsText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for DefensiveReboundsText to Int - %w; defaulting to 0.", DefensiveReboundsText, err))
 						boxScoreStats.DefensiveRebounds = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					TotalReboundsText := util.CleanTextDatum(s.Find("td:nth-child(14)").Text())
 					boxScoreStats.TotalRebounds, err = util.TextToInt(TotalReboundsText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for TotalReboundsText to Int\n", TotalReboundsText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for TotalReboundsText to Int - %w; defaulting to 0.", TotalReboundsText, err))
 						boxScoreStats.TotalRebounds = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					AssistsText := util.CleanTextDatum(s.Find("td:nth-child(15)").Text())
 					boxScoreStats.Assists, err = util.TextToInt(AssistsText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for AssistsText to Int\n", AssistsText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for AssistsText to Int - %w; defaulting to 0.", AssistsText, err))
 						boxScoreStats.Assists = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					StealsText := util.CleanTextDatum(s.Find("td:nth-child(16)").Text())
 					boxScoreStats.Steals, err = util.TextToInt(StealsText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for StealsText to Int\n", StealsText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for StealsText to Int - %w; defaulting to 0.", StealsText, err))
 						boxScoreStats.Steals = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					BlocksText := util.CleanTextDatum(s.Find("td:nth-child(17)").Text())
 					boxScoreStats.Blocks, err = util.TextToInt(BlocksText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for BlocksText to Int\n", BlocksText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for BlocksText to Int - %w; defaulting to 0.", BlocksText, err))
 						boxScoreStats.Blocks = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					TurnoversText := util.CleanTextDatum(s.Find("td:nth-child(18)").Text())
 					boxScoreStats.Turnovers, err = util.TextToInt(TurnoversText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for TurnoversText to Int\n", TurnoversText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for TurnoversText to In - %w; defaulting to 0.", TurnoversText, err))
 						boxScoreStats.Turnovers = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					PersonalFoulsText := util.CleanTextDatum(s.Find("td:nth-child(19)").Text())
 					boxScoreStats.PersonalFouls, err = util.TextToInt(PersonalFoulsText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for PersonalFoulsText to Int\n", PersonalFoulsText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for PersonalFoulsText to Int - %w; defaulting to 0.", PersonalFoulsText, err))
 						boxScoreStats.PersonalFouls = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					PointsText := util.CleanTextDatum(s.Find("td:nth-child(20)").Text())
 					boxScoreStats.Points, err = util.TextToInt(PointsText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for PointsText to Int\n", PointsText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for PointsText to Int - %w; defaulting to 0.", PointsText, err))
 						boxScoreStats.Points = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					GameScoreText := util.CleanTextDatum(s.Find("td:nth-child(21)").Text())
 					boxScoreStats.GameScore, err = util.TextToFloat32(GameScoreText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for GameScoreText to Float64\n", GameScoreText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for GameScoreText to Float64 - %w; defaulting to 0.", GameScoreText, err))
 						boxScoreStats.GameScore = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 
 					PlusMinusText := util.CleanTextDatum(s.Find("td:nth-child(22)").Text())
 					boxScoreStats.PlusMinus, err = util.TextToInt(PlusMinusText)
 					if err != nil {
-						log.Printf("Cannot convert '%s' for PlusMinusText to Int\n", PlusMinusText)
+						log.Println(fmt.Errorf("WARNING: Cannot convert '%s' for PlusMinusText to Int - %w; defaulting to 0.", PlusMinusText, err))
 						boxScoreStats.PlusMinus = 0
-						log.Printf("WARNING: %s\n", err.Error())
 					}
 				} else {
 					boxScoreStats.MinutesPlayed = 0
 				}
 				basicNBABoxScoreStats = append(basicNBABoxScoreStats, boxScoreStats)
 			} else {
-				s.Find(basicBoxScoreReserveHeaders).Each(func(_ int, s *goquery.Selection) {
+				s.Find(boxScoreReserveHeaders).Each(func(_ int, s *goquery.Selection) {
 
 					reserveHeader = util.CleanTextDatum(s.Text())
 					_, ok := basicBoxScoreReservesHeaderValues[reserveHeader]